server/app/requirements.txt
fastapi>=0.113.0,<0.114.0
uvicorn[standard]>=0.30.0
numpy<2
pillow
opencv-python-headless

